+++ src/Image/GdalRaster.cpp
@@ -1274,7 +1274,7 @@ namespace {
 				val = 1.0f;
 			if( biRange > 1.0f )
 			{
-				result.right = ( pow(val - lowVal, 2)/pow(biRange, 2) );
+				result.right = ( powf(val - lowVal, 2) / powf(biRange, 2) );
 				result.left = 1.0f - result.right;
 			}					
 			else
